pyproject.toml
[project]
name = "f1-penalty-agent"
version = "0.1.0"
description = "AI agent for explaining F1 penalties and FIA regulations to F1 viewers"
authors = [
    {name = "addis"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"

dependencies = [
    # Data collection
    "fastf1>=3.7.0",
    "requests>=2.32.0",
    "beautifulsoup4>=4.14.0",
    "pypdf>=6.4.0",
    "lxml>=6.0.0",

    # RAG pipeline
    "sentence-transformers>=5.1.0",  # Updated from 3.0.0
    "qdrant-client>=1.16.0",  # Updated from 1.12.0

    # LLM
    "google-genai>=1.56.0",  # Updated from 1.0.0

    # Utilities
    "python-dotenv>=1.0.0",
    "pydantic>=2.12.0",  # Updated from 2.11.0
    "pydantic-settings>=2.7.0",
    "rich>=13.9.0",
    "typer>=0.21.0",  # Updated from 0.15.0

    # Web API
    "fastapi>=0.128.0",  # Updated from 0.115.0
    "uvicorn[standard]>=0.40.0",  # Updated from 0.34.0
]

[project.scripts]
f1agent = "src.adapters.inbound.cli:app"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{include = "src"}]

[tool.poetry.group.dev.dependencies]
pytest = ">=9.0.0"
pytest-asyncio = ">=1.3.0"
ruff = ">=0.14.0"
pre-commit = "^4.5.1"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]  # Allow long lines in prompt strings

[tool.mypy]
python_version = "3.12"
strict = true

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
